// src/Avalanche/utils/types.ts
export type DexType = 
  | "uniswapv3" 
  | "lfjv22"
  | "lfj"
  | "lfjv21"